refactor: retire thin repository wrappers
