--- backend/app/repositories/course_repository.py
@@ -1,8 +1,0 @@
-from sqlalchemy.orm import Session
-
-from ..models import CourseMood
-from ..repository_normalized import list_courses
-
-
-def list_course_entries(db: Session, mood: CourseMood | None):
-    return list_courses(db, mood)
