+++ backend/app/services/course_service.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 
 from ..models import CourseMood
-from ..repositories.course_repository import list_course_entries
+from ..repository_normalized import list_courses as list_course_entries
 
 
 def read_courses_service(db: Session, mood: CourseMood | None):
